backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.6
google-cloud-speech==2.19.0
google-cloud-texttospeech==2.14.1
openai==0.27.6
websockets==11.0.3 